+++ arrayfire/src/api/c/gaussian_kernel.cpp
@@ -47,7 +47,7 @@ Array<T> gaussianKernel(const int rows, const int cols, const double sigma_r,
         Array<T> w  = transpose<T>(wt, false);
 
         Array<T> c = createValueArray<T>(
-            odims, scalar<T>(static_cast<double>(cols - 1) / 2.0));
+            odims, scalar<T>(static_cast<float>(cols - 1) / 2.0));
         w = arithOp<T, af_sub_t>(w, c, odims);
 
         sigma        = sigma_c > 0 ? sigma_c : 0.25 * cols;
@@ -62,7 +62,7 @@ Array<T> gaussianKernel(const int rows, const int cols, const double sigma_r,
         Array<T> w = range<T>(dim4(rows, cols), 0);
 
         Array<T> r = createValueArray<T>(
-            odims, scalar<T>(static_cast<double>(rows - 1) / 2.0));
+            odims, scalar<T>(static_cast<float>(rows - 1) / 2.0));
         w = arithOp<T, af_sub_t>(w, r, odims);
 
         sigma        = sigma_r > 0 ? sigma_r : 0.25 * rows;
